backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
PyPDF2==3.0.1
werkzeug==3.0.1
email-validator==2.1.0
scikit-learn==1.3.2
